# Use bucket sort for token logits