PR https://github.com/ggerganov/llama.cpp/pull/5085 attempts to speed up token sampling for cases with large --top-k by using `std::nth_element` and then `std::sort`. This PR instead exploits the limited range of token logits to implement bucket sort. There are 100 equal-sized buckets in the logit range [-10, 10]. Tokens are first distributed to these buckets. Then only those buckets relevant for --top-k are actually sorted. This is also faster for --top-k 32000 because the runtime for comparison-based sorting algorithms increases more than linearly with the number of elements. I am assuming that the top tokens beyond --top-k can be arbitrarily changed without affecting program correctness.

For LLaMA 2 7b q8_0 I get on my system with an RTX 3090 and a Ryzen 3700X:

| --top-k | t/s master | t/s nth_element | Speedup nth_element | t/s bucket sort | Speedup bucket_sort |
|---------|------------|-----------------|---------------------|-----------------|---------------------|
|    1000 |       1991 |            1973 |                0.99 |            2213 |                1.11 |
|    2000 |       1396 |            1375 |                0.98 |            2065 |                1.48 |
|    4000 |        893 |            1478 |                1.66 |            1821 |                2.04 |
|    8000 |        574 |            1123 |                1.96 |            1416 |                2.47 |
|   16000 |        389 |             778 |                2.00 |             997 |                2.56 |
|   31780 |        309 |             492 |                1.59 |             639 |                2.07 |
|   32000 |        492 |             485 |                0.99 |             637 |                1.29 |

![bucket_sort](https://github.com/ggerganov/llama.cpp/assets/18492268/40489474-ac7a-45bd-897a-91ef19b0e7e9)