@ikawrakow 2024-01-24
I made a stand-alone benchmark (attached). It reads logits from a file (previously generated with `perplexity --kl-divergence-base` and tests `std::sort/std::partial_sort` against this bucket sort. On my system the standard library is faster than this bucket sort by 2X..20X, depending on `top_k`. Which is not surprising, given the few hundred allocations that happen in the bucket sort.

For instance
```
./perplexity -m models/Mistral-7B/ggml-model-f16.gguf -f tests/wiki.test.raw --kl-divergence-base m7.logits

g++ -O3 -o test_sorting test_sorting.cpp

./test_sorting m7.logits 100 50
std::sort:
  <time> = 21.241 us
  top_1  = 0.700492
  top_k  = 0.000402343
Bucket:
  <time> = 437.53 us
  top_1  = 0.700492
  top_k  = 0.000402343

./test_sorting some_logits 100 500
std::sort:
  <time> = 75.9663 us
  top_1  = 0.700492
  top_k  = 1.52254e-05
Bucket:
  <time> = 504.853 us
  top_1  = 0.700492
  top_k  = 1.52254e-05

./test_sorting some_logits 100 5000
std::sort:
  <time> = 236.347 us
  top_1  = 0.700492
  top_k  = 1.92905e-07
Bucket:
  <time> = 664.981 us
  top_1  = 0.700492
  top_k  = 1.92905e-07
```

[test_sorting.cpp.gz](https://github.com/ggerganov/llama.cpp/files/14033380/test_sorting.cpp.gz)