@JohannesGaessler 2024-01-24
With the benchmark code you posted I also get worse performance than with the command I used to test my implementation, e.g. `./main --model models/nvme/${model_name}-${quantization}.gguf -ngl 99 --ctx-size 4096 --ignore-eos --n-predict 256 --seed 1337 --top-k 32000`. But I think the issue is that the benchmark is simply not representative in terms of logit value distribution since you do some transformations which - for the calculation of KL divergence - arenegligible. I added a simple debug print to inspect how the values are being distributed to buckets with --top-k 32000 and with main you end up with buckets holding up to ~3% of the values while with the benchmark a single bucket holds ~80% of the values.

<details>
<summary>main</details>

```
ib=99 buckets[ib].size()=46
ib=98 buckets[ib].size()=4
ib=97 buckets[ib].size()=5
ib=96 buckets[ib].size()=4
ib=95 buckets[ib].size()=8
ib=94 buckets[ib].size()=6
ib=93 buckets[ib].size()=9
ib=92 buckets[ib].size()=10
ib=91 buckets[ib].size()=14
ib=90 buckets[ib].size()=17
ib=89 buckets[ib].size()=17
ib=88 buckets[ib].size()=19
ib=87 buckets[ib].size()=22
ib=86 buckets[ib].size()=31
ib=85 buckets[ib].size()=20
ib=84 buckets[ib].size()=27
ib=83 buckets[ib].size()=35
ib=82 buckets[ib].size()=36
ib=81 buckets[ib].size()=32
ib=80 buckets[ib].size()=44
ib=79 buckets[ib].size()=48
ib=78 buckets[ib].size()=60
ib=77 buckets[ib].size()=68
ib=76 buckets[ib].size()=56
ib=75 buckets[ib].size()=93
ib=74 buckets[ib].size()=91
ib=73 buckets[ib].size()=95
ib=72 buckets[ib].size()=110
ib=71 buckets[ib].size()=130
ib=70 buckets[ib].size()=158
ib=69 buckets[ib].size()=157
ib=68 buckets[ib].size()=162
ib=67 buckets[ib].size()=207
ib=66 buckets[ib].size()=194
ib=65 buckets[ib].size()=269
ib=64 buckets[ib].size()=302
ib=63 buckets[ib].size()=329
ib=62 buckets[ib].size()=371
ib=61 buckets[ib].size()=363
ib=60 buckets[ib].size()=445
ib=59 buckets[ib].size()=452
ib=58 buckets[ib].size()=518
ib=57 buckets[ib].size()=556
ib=56 buckets[ib].size()=605
ib=55 buckets[ib].size()=672
ib=54 buckets[ib].size()=738
ib=53 buckets[ib].size()=838
ib=52 buckets[ib].size()=834
ib=51 buckets[ib].size()=897
ib=50 buckets[ib].size()=994
ib=49 buckets[ib].size()=962
ib=48 buckets[ib].size()=996
ib=47 buckets[ib].size()=1075
ib=46 buckets[ib].size()=1068
ib=45 buckets[ib].size()=1110
ib=44 buckets[ib].size()=1120
ib=43 buckets[ib].size()=1091
ib=42 buckets[ib].size()=1136
ib=41 buckets[ib].size()=1126
ib=40 buckets[ib].size()=1126
ib=39 buckets[ib].size()=950
ib=38 buckets[ib].size()=1046
ib=37 buckets[ib].size()=921
ib=36 buckets[ib].size()=825
ib=35 buckets[ib].size()=824
ib=34 buckets[ib].size()=753
ib=33 buckets[ib].size()=621
ib=32 buckets[ib].size()=679
ib=31 buckets[ib].size()=552
ib=30 buckets[ib].size()=456
ib=29 buckets[ib].size()=414
ib=28 buckets[ib].size()=324
ib=27 buckets[ib].size()=269
ib=26 buckets[ib].size()=258
ib=25 buckets[ib].size()=219
ib=24 buckets[ib].size()=184
ib=23 buckets[ib].size()=131
ib=22 buckets[ib].size()=94
ib=21 buckets[ib].size()=100
ib=20 buckets[ib].size()=81
ib=19 buckets[ib].size()=64
ib=18 buckets[ib].size()=49
ib=17 buckets[ib].size()=38
ib=16 buckets[ib].size()=30
ib=15 buckets[ib].size()=24
ib=14 buckets[ib].size()=18
ib=13 buckets[ib].size()=10
ib=12 buckets[ib].size()=12
ib=11 buckets[ib].size()=6
ib=10 buckets[ib].size()=5
ib=9 buckets[ib].size()=3
ib=8 buckets[ib].size()=3
ib=7 buckets[ib].size()=2
ib=6 buckets[ib].size()=3
ib=5 buckets[ib].size()=1
ib=4 buckets[ib].size()=0
ib=3 buckets[ib].size()=1
ib=2 buckets[ib].size()=0
ib=1 buckets[ib].size()=0
ib=0 buckets[ib].size()=2
```

</details>


<details>
<summary>test_sorting</details>

```
ib=99 buckets[ib].size()=0
ib=98 buckets[ib].size()=0
ib=97 buckets[ib].size()=0
ib=96 buckets[ib].size()=1
ib=95 buckets[ib].size()=0
ib=94 buckets[ib].size()=0
ib=93 buckets[ib].size()=0
ib=92 buckets[ib].size()=0
ib=91 buckets[ib].size()=1
ib=90 buckets[ib].size()=1
ib=89 buckets[ib].size()=0
ib=88 buckets[ib].size()=0
ib=87 buckets[ib].size()=0
ib=86 buckets[ib].size()=0
ib=85 buckets[ib].size()=0
ib=84 buckets[ib].size()=0
ib=83 buckets[ib].size()=1
ib=82 buckets[ib].size()=1
ib=81 buckets[ib].size()=1
ib=80 buckets[ib].size()=0
ib=79 buckets[ib].size()=0
ib=78 buckets[ib].size()=1
ib=77 buckets[ib].size()=2
ib=76 buckets[ib].size()=0
ib=75 buckets[ib].size()=1
ib=74 buckets[ib].size()=0
ib=73 buckets[ib].size()=2
ib=72 buckets[ib].size()=2
ib=71 buckets[ib].size()=1
ib=70 buckets[ib].size()=3
ib=69 buckets[ib].size()=2
ib=68 buckets[ib].size()=1
ib=67 buckets[ib].size()=2
ib=66 buckets[ib].size()=3
ib=65 buckets[ib].size()=4
ib=64 buckets[ib].size()=4
ib=63 buckets[ib].size()=6
ib=62 buckets[ib].size()=8
ib=61 buckets[ib].size()=6
ib=60 buckets[ib].size()=5
ib=59 buckets[ib].size()=6
ib=58 buckets[ib].size()=4
ib=57 buckets[ib].size()=3
ib=56 buckets[ib].size()=6
ib=55 buckets[ib].size()=11
ib=54 buckets[ib].size()=15
ib=53 buckets[ib].size()=14
ib=52 buckets[ib].size()=10
ib=51 buckets[ib].size()=20
ib=50 buckets[ib].size()=20
ib=49 buckets[ib].size()=29
ib=48 buckets[ib].size()=26
ib=47 buckets[ib].size()=25
ib=46 buckets[ib].size()=27
ib=45 buckets[ib].size()=33
ib=44 buckets[ib].size()=49
ib=43 buckets[ib].size()=48
ib=42 buckets[ib].size()=51
ib=41 buckets[ib].size()=52
ib=40 buckets[ib].size()=73
ib=39 buckets[ib].size()=68
ib=38 buckets[ib].size()=92
ib=37 buckets[ib].size()=100
ib=36 buckets[ib].size()=102
ib=35 buckets[ib].size()=106
ib=34 buckets[ib].size()=119
ib=33 buckets[ib].size()=147
ib=32 buckets[ib].size()=150
ib=31 buckets[ib].size()=191
ib=30 buckets[ib].size()=200
ib=29 buckets[ib].size()=201
ib=28 buckets[ib].size()=241
ib=27 buckets[ib].size()=285
ib=26 buckets[ib].size()=282
ib=25 buckets[ib].size()=336
ib=24 buckets[ib].size()=345
ib=23 buckets[ib].size()=378
ib=22 buckets[ib].size()=409
ib=21 buckets[ib].size()=446
ib=20 buckets[ib].size()=468
ib=19 buckets[ib].size()=512
ib=18 buckets[ib].size()=597
ib=17 buckets[ib].size()=557
ib=16 buckets[ib].size()=25087
```

Buckets below `ib=16` are empty.

</details>