@ikawrakow 2024-01-24
In which buckets the tokens end up is a matter of context. You see in the example you posted above that many logits ended up in bucket 16. As nothing was done to the logits above bucket 16, this means that without the processing done in `perplexity` when saving the logit data, the 25k+ tokens that are now all in bucket 16 would have been distributed in just 16 buckets rather than 60+ buckets as it happens in your usage. 

But if for the sake of argument we assume for a moment that the test case I made is not realistic, the PR still lowers performance for more normal usage with `top_k < 1000`. On my system with empty prompt and 256 generated tokens I get

| top_k | Master t/s | PR t/s | Speedup |
|---:|---:|---:|---:|
|40 |  8648 | 3034 | 0.351 |
|100 |7758 | 3076 | 0.396 |
|200 | 6796 |3039 | 0.447 |
|500 | 4847 | 2905 | 0.599 |
|1000 | 3353 | 2872 | 0.856 |
|2000 | 2133 | 2592 | 1.215 |

Where the crossover to the STL having better performance than your bucket sort very likely dependent on many details (e.g., compiler, compiler version, CPU, context). So, I think, you definitely need a check there for `k > some_constant` where the bucket sort is being applied.